pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "ghost-in-the-droid"
version = "1.0.0"
description = "Give any AI agent an Android body. Open-source phone automation via ADB — tap, swipe, type, run skills, scale across devices."
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["android", "adb", "automation", "ghost", "agent", "mcp", "fastapi", "phone-farm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "psutil>=5.9",
    "openai>=1.0",
]

[project.optional-dependencies]
llm = ["anthropic>=0.30"]
analytics = ["google-analytics-data>=0.18"]
ocr = ["rapidocr-onnxruntime>=1.3"]
test = ["playwright>=1.40", "pytest>=7.0"]
all = ["ghost-in-the-droid[llm,analytics,ocr,test]"]

[project.scripts]
android-agent = "gitd.cli:main"

[tool.setuptools.packages.find]
include = ["gitd*"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.format]
quote-style = "double"
